backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.2.0

# Data Validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Environment Variables
python-dotenv==1.0.0

# Background Tasks & Scheduling
schedule==1.2.0

# Utilities
python-dateutil==2.8.2